Fix details

+++ src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/resultDTOs/PoliticalResult.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -19,5 +21,5 @@ public class PoliticalResult {
     private int votes;
 
     @JsonProperty("porcentaje")
-    private Double percentage;
+    private BigDecimal percentage;
 }
